% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRdFileNames.R
\name{getRdFileNames}
\alias{getRdFileNames}
\title{Track Rd file names at which 'topic' is documented}
\usage{
getRdFileNames(topic, package = NULL)
}
\arguments{
\item{topic}{A length-one character vector specifying the topic (alias).}

\item{package}{A character vector given the packages to search for Rd file
names that document the \code{topic} , or 'NULL'. By default, all the
packages in the search path are used.}
}
\description{
Tracks the Rd file names at which a given 'topic' (alias) is documented.
}
\examples{

getRdFileNames("rbind")

isInstalled <- function(pkg)
  inherits(suppressWarnings(packageDescription(pkg)), "packageDescription")

if (isInstalled("IRanges"))
  getRdFileNames("rbind", package=c("base", "IRanges"))

if (isInstalled("Biobase"))
  getRdFileNames("ExpressionSet", "Biobase")

}
\author{
Chao-Jen Wong \email{cwon2@fhcrc.org}
}
\keyword{programming}
