% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findExternalDeps.R
\name{findExternalDeps}
\alias{findExternalDeps}
\title{Find External Dependencies of a Package}
\usage{
findExternalDeps(package)
}
\arguments{
\item{package}{the quoted name of the package to analyze.}
}
\value{
A list with four components: \code{S4Classes}, \code{S4Methods},
\code{functions}, and \code{variables}. Each of these components in turn
holds a named list of character vectors of object names from external
packages.
}
\description{
Finds external dependencies of a package.
}
\details{
The result is an approximation based upon the findings of \code{findGlobals}
and an exploration of the package namespace.
}
\examples{

library(stats4)
findExternalDeps("stats4")
}
\seealso{
\code{\link[codetools]{findGlobals}},
%\code{\link[codetools]{checkUsagePackage}}, Rd file not found.
}
\author{
Patrick Aboyoun
}
\keyword{programming}
